refactor(api): types
