+++ crates/api/src/user/get_user.rs
@@ -1,48 +1,7 @@
-use crate::{league::League, HttpClient, HttpClientError};
 use reqwest::Method;
-use serde_json::Value;
 
-#[derive(Debug)]
-pub struct User {
-  pub name: String,
-  pub id: String,
-  pub image: String,
-  pub leagues: Vec<League>,
-}
-
-impl From<Value> for User {
-  fn from(value: Value) -> Self {
-    let user = value.get("user").unwrap();
-
-    let id = user
-      .get("id")
-      .unwrap()
-      .as_str()
-      .unwrap_or_default()
-      .to_string();
-
-    let name = user
-      .get("name")
-      .unwrap()
-      .as_str()
-      .unwrap_or_default()
-      .to_string();
-
-    let image = user
-      .get("profile")
-      .map(|v| v.as_str().unwrap_or_default().to_string())
-      .unwrap_or_default();
-
-    let leagues = vec![];
-
-    Self {
-      id,
-      name,
-      image,
-      leagues,
-    }
-  }
-}
+use super::User;
+use crate::{HttpClient, HttpClientError};
 
 impl HttpClient {
   pub async fn get_user(&self) -> Result<User, HttpClientError> {
