+++ crates/api/src/user/league_user_info.rs
@@ -8,7 +8,7 @@ impl HttpClient {
     league_id: &str,
   ) -> Result<HttpResponse<Value>, HttpClientError> {
     let url = format!("/leagues/{}/me", league_id);
-    let response = self.get::<Value>(Method::GET, &url).await?;
+    let response = self.get(Method::GET, &url).await?;
     Ok(response)
   }
 }
